// src/interfaces/projects.ts
// Generated by https://quicktype.io

export interface ProjectCollection {
    projectCollection: ProjectCollectionClass;
}

export interface ProjectCollectionClass {
    limit?: number;
    total?: number;
    items?: Item[];
}

export interface Item {
    title:          string;
    description:    string;
    receivedAGrant: boolean;
    isComingSoon:   boolean;
    image:          Image;
}

export interface Image {
    title: string;
    url:   string;
}